requirements.txt
apache-airflow>=2.5.0
pyspark>=3.3.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
openpyxl>=3.0.0
lxml>=4.9.0
numpy>=1.21.0
pyarrow>=10.0.0